#76 Move uts-run runner code to separate module
